requirements.txt
django-qsstats-magic-0.7.1 #django-qsstats-magic, in installed apps as qsstats
south   #
taggit  #
